requirements.txt
efficientnet-pytorch==0.5.1
entrypoints==0.3
flow-vis==0.1
matplotlib==3.1.3
mlflow==1.6.0
numpy==1.17.4
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.32
Pillow==7.0.0
scikit-learn==0.22.1
scipy==1.4.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1